(PUP-6453) Test using in module data to avoid ordering issues
PUP-6453 prevents us from using class params of classes included in a
component in that components produced resources. This is a workaround
showing how you can use in module data to accomplish the same thing.

Instead of referencing the class param in the produce statement you can
use lookup to lookup the value of that class param. This won't work if a
user overrides the bound value of that param with an ENC or resource
like declaration.